+++ src/main/java/chess/dao/PieceDao.java
@@ -1,6 +1,6 @@
 package chess.dao;
 
-import chess.dto.request.PieceDto;
+import chess.dto.PieceDto;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
